All the required Examples for this Project
Note - I had some problem in running the mp3dec sketch to burn the mp3 decoder for Audio applications. So, to make sure you don't face the same, I have added the MP3DEC File, so just create a BIN folder in your sd card root, and paste this file into it, and you are done!

Another thing, if your device cant detect the SD card, just press down the Spresense main Board a little, as there is a somewhat loose connection between the main and extension board for me.